fix: fix type of `zoomLevel` * change type of `zoomLevel` from int to double

+++ lib/src/interface_osm/osm_interface.dart
@@ -165,7 +165,7 @@ abstract class OSMPlatform extends PlatformInterface {
 
   Future<void> setZoom(
     int idOSM, {
-    int? zoomLevel,
+    double? zoomLevel,
     double? stepZoom,
   });
